+++ examples/validator/main.go
@@ -20,25 +20,22 @@ var (
 
 func main() {
     validator := mrview.NewValidator()
-    err := validator.Register("login", ValidateLogin)
 
-    if err != nil {
+    if err := validator.Register("login", ValidateLogin); err != nil {
         fmt.Println(err)
         return
     }
 
     user := User{Login: "valid-login"}
-    err = validator.Validate(context.Background(), &user)
 
-    if err != nil {
+    if err := validator.Validate(context.Background(), &user); err != nil {
         fmt.Println(err)
         return
     }
 
     user2 := User{Login: "not-valid-login!"}
-    err = validator.Validate(context.Background(), &user2)
 
-    if err != nil {
+    if err := validator.Validate(context.Background(), &user2); err != nil {
         fmt.Println(err)
         return
     }